requirements/base.txt
# Django
Django>=3.1.3,<3.2.0
Pillow>=8.0.1,<8.1.0
argon2-cffi>=20.1.0,<21.0.0

# Environ
django-environ==0.4.5,<0.5.0

# CORS Headers
django_cors_headers>=3.5.0,<3.6.0

# GraphQL
graphene_django>=2.13.0,<2.14.0

# JWT
django_graphql_jwt>=0.3.1,<0.4.0 # https://github.com/flavors/django-graphql-jwt
